web report changes

+++ Ginger/GingerCoreNET/Logger/WebReportGenerator.cs
@@ -61,7 +61,7 @@ namespace Amdocs.Ginger.CoreNET.Logger
 
             try
             {
-                json = $"window.runsetData={json}";
+                json = $"window.runsetData={json};";
                 StringBuilder pageDataSb = new StringBuilder();
                 pageDataSb.Append("file:///");
                 pageDataSb.Append(clientAppFolderPath.Replace('\\', '/'));
@@ -155,8 +155,11 @@ namespace Amdocs.Ginger.CoreNET.Logger
                             {
                                 string fileName = Path.GetFileName(screenshot);
                                 string newScreenshotPath = Path.Combine(imageFolderPath, fileName);
-                                System.IO.File.Copy(screenshot, newScreenshotPath, true); //TODO - Replace with the real location under Ginger installation
-                                newScreenShotsList.Add(fileName);
+                                if (File.Exists(newScreenshotPath))
+                                {
+                                    System.IO.File.Copy(screenshot, newScreenshotPath, true); //TODO - Replace with the real location under Ginger installation
+                                    newScreenShotsList.Add(fileName);
+                                }
                             }
                             liteDbAction.ScreenShots = newScreenShotsList;
                         }
